pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "texfury"
version = "1.0.0"
description = "Fast image-to-DDS conversion and GTA V YTD texture dictionaries"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
pil = ["Pillow"]
